fix(vizia): font typeface, clipboard, knob drag, shared-ui button unification
Root-cause fixes affecting all 6 plugins:
- vg::Font::default() had no typeface attached, draw_str() silently drew
  nothing with it - this was the cause of every "dB numbers missing"
  report (peak meters, goniometer correlation, EQ grid labels). Now loads
  a real system typeface once via FontMgr, cached in shared-ui/canvas.rs.
- vizia's clipboard feature was never enabled (default-features=false,
  features=["baseview"] only) - paste never worked anywhere. Added
  "clipboard" feature to all 6 plugin Cargo.toml, pulls in copypasta.
- KnobView/HSliderView drag could get stuck if a MouseUp was missed
  (e.g. host stealing focus mid-drag) - dragging stayed true forever,
  ignoring input until a fresh click. Now checks the actual button state
  on every MouseMove and self-heals.

shared-ui unification (was fragmented per plugin, DeepSeek flagged it):
- New shared-ui/src/buttons.rs: BUTTON_HEIGHT/_SMALL, KNOB_SIZE,
  SLIDER_HEIGHT constants + AMBER/IDLE_BG/DANGER_BG/DANGER_TEXT colors +
  toggle_button()/toggle_button_small()/danger_button() helpers - the old
  Iced toggle_button/output_tools_strip pattern that never got ported to
  Vizia, root cause of several inconsistent/dead toggle buttons.
- Aether, Equilibrium, Meridian, Lucent migrated onto the shared helpers
  (Bypass, Mono/Delta/Solo/Listen/Pre-Master, EQ band-type, slope
  selectors, Split/Clip, Crossfeed realism, Standalone/Relay/Hybrid,
  Resonance/relay toggles, all RESET buttons). Aurum intentionally left
  on its own local kn()/tog()/cyc() - UI still WIP.

Per-plugin fixes:
- Aether: SAVE silently no-op'd with no vault path configured - now
  falls back to the plugin data dir like Meridian/Equilibrium. Vault
  setup screen rebuilt to match Lucent's card layout. EQ curve fill
  removed (shared SpectrumView eq_curve rendering). Textbox/ComboBox
  sizes reduced.
- Equilibrium: goniometer correlation dot colors restored to the
  original Iced values (had drifted during the port). Footer analyze
  section and preset list alignment fixed. Toggle/reset button heights
  reduced.
- Meridian: reset button only touched DSP, not the UI widgets - top
  strip and footer knob sections now rebuild on params_gen like
  Equilibrium already did. Width param's right-click reset used the
  wrong default (0% instead of 100%). Window height increased for
  header/footer breathing room. Peak meter extended toward the
  goniometer.
- Lucent: Sensitivity knob was fed a raw 0-100 value where KnobView
  expects a normalized 0-1 value_norm, spun into a garbled sunburst.
  Masking toggle never reacted to AnalyzeMode changes after the initial
  build. Mode switch button recolored amber. Duplicate "LX AUDIOLABS"
  sidebar label removed. Relay bar left-aligned.
- Lucent-Relay: window was 260x160, too small for the target ComboBox's
  popup to render without clipping (vizia's Popup has should_reposition
  disabled) - grown to 260x240. Textbox/ComboBox sizes reduced.

+++ plugins/aurum/src/lib.rs
@@ -370,18 +370,18 @@ impl PluginLogic for Aurum {
             let in_l = buffer.input(0)[i];
             let in_r = buffer.input(1)[i];
 
-            let clip_ceiling = 10.0_f32.powf(self.params.clip_ceiling.value() as f32 / 20.0);
-            let clip_softness = self.params.clip_softness.value() as f32 / 100.0;
-            let sat_drive_stereo = self.params.sat_drive_stereo.value() as f32;
-            let sat_drive_mid = self.params.sat_drive_mid.value() as f32;
-            let sat_drive_side = self.params.sat_drive_side.value() as f32;
-            let sat_mix = self.params.sat_mix.value() as f32 / 100.0;
-            let width = self.params.stereo_width.value() as f32;
-            let mb_global_gain = 10.0_f32.powf(self.params.mb_global_gain.value() as f32 / 20.0);
-            let mb_makeup_lo = 10.0_f32.powf(self.params.mb_gain_mid_lo.value() as f32 / 20.0);
-            let mb_makeup_hi = 10.0_f32.powf(self.params.mb_gain_mid_hi.value() as f32 / 20.0);
-            let mb_makeup_side = 10.0_f32.powf(self.params.mb_gain_side.value() as f32 / 20.0);
-            let out_gain = 10.0_f32.powf(self.params.output_gain.value() as f32 / 20.0);
+            let clip_ceiling = 10.0_f32.powf(self.params.clip_ceiling.value() / 20.0);
+            let clip_softness = self.params.clip_softness.value() / 100.0;
+            let sat_drive_stereo = self.params.sat_drive_stereo.value();
+            let sat_drive_mid = self.params.sat_drive_mid.value();
+            let sat_drive_side = self.params.sat_drive_side.value();
+            let sat_mix = self.params.sat_mix.value() / 100.0;
+            let width = self.params.stereo_width.value();
+            let mb_global_gain = 10.0_f32.powf(self.params.mb_global_gain.value() / 20.0);
+            let mb_makeup_lo = 10.0_f32.powf(self.params.mb_gain_mid_lo.value() / 20.0);
+            let mb_makeup_hi = 10.0_f32.powf(self.params.mb_gain_mid_hi.value() / 20.0);
+            let mb_makeup_side = 10.0_f32.powf(self.params.mb_gain_side.value() / 20.0);
+            let out_gain = 10.0_f32.powf(self.params.output_gain.value() / 20.0);
 
             // [2] Clipper
             let (cl, cr) = if clip_ms_mode {
